refactor: restructure codebase into layered architecture
Major reorganization of the flat ts/ directory into a layered structure:
- core/ (types, constants, van.ts)
- state/ (UIState + DataState, split from monolithic AppState)
- services/ (storage, tree, favicon, theme, validator, new CommandService)
- components/ (tree, edit-form, settings, footer)

Key improvements:
- CommandService decouples UI from persistence layer
- Deduplicated save-and-rebuild-tree pattern (was repeated 5+ times)
- Extracted form reset into reusable function
- Removed dead code (createdTable state, commented-out code)
- Moved SVG icons from inline strings in constants.ts to .svg files
- Split monolithic custom.css (571 lines) into 8 focused CSS files
- Fixed dark mode hover bug in tree list (hardcoded lightgray → var)

Storage-persisted enum values are preserved to avoid breaking existing users.
