+++ static/css/base.css
@@ -1,0 +1,18 @@
+/* ===== BASE STYLES ===== */
+body {
+  color: var(--text-color);
+  background-color: var(--background-color);
+  font-size: 14px;
+  margin-left: 0;
+  margin-right: 0;
+}
+
+a {
+  text-decoration: none;
+  color: var(--link-color);
+}
+
+/* ===== UTILITY CLASSES ===== */
+.display-none {
+  display: none;
+}
