+++ static/css/forms.css
@@ -1,0 +1,88 @@
+/* ===== FORM LAYOUT ===== */
+.form-header {
+  font-size: 16px;
+  font-weight: 500;
+  margin-bottom: 15px;
+  color: var(--text-color);
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding-bottom: 10px;
+  border-bottom: 1px solid var(--border-color);
+}
+
+.form-actions {
+  display: flex;
+  gap: 10px;
+  margin-top: 10px;
+}
+
+.form-hint {
+  margin-top: 15px;
+  font-size: 12px;
+  color: var(--footer-link-color);
+  font-style: italic;
+  padding: 5px;
+  border-top: 1px solid var(--border-color);
+}
+
+/* ===== FORM INPUTS ===== */
+#newlink-form input[type="text"] {
+  width: 100%;
+  padding: 8px 10px;
+  border: 1px solid var(--border-color);
+  border-radius: 4px;
+  margin-bottom: 10px;
+  background-color: var(--background-color);
+  color: var(--text-color);
+  box-sizing: border-box;
+}
+
+#newlink-form input[type="submit"],
+#newlink-form input[type="button"] {
+  padding: 6px 12px;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  transition: background-color 0.2s;
+}
+
+#newlink-form input[type="submit"] {
+  background-color: transparent;
+  color: var(--primary-button-color);
+  border: 1px solid var(--primary-button-color);
+  font-weight: 500;
+}
+
+#newlink-form input[type="submit"]:hover {
+  background-color: rgba(66, 133, 244, 0.1);
+  color: var(--primary-button-hover-color);
+}
+
+#newlink-form input[type="button"] {
+  background-color: var(--hover-bg-color);
+  color: var(--text-color);
+}
+
+#newlink-form input[type="button"]:hover {
+  background-color: var(--border-color);
+}
+
+/* ===== SELECT DROPDOWNS ===== */
+.select-wrapper {
+  position: relative;
+  width: 100%;
+  margin-bottom: 10px;
+}
+
+.select-wrapper select {
+  width: 100%;
+  padding: 8px 10px;
+  border: 1px solid var(--border-color);
+  border-radius: 4px;
+  background-color: var(--background-color);
+  color: var(--text-color);
+  box-sizing: border-box;
+  appearance: none;
+  cursor: pointer;
+}
