+++ static/css/layout.css
@@ -1,0 +1,23 @@
+/* ===== LAYOUT ===== */
+.row {
+  display: flex;
+  min-height: calc(100vh - 80px);
+}
+
+.main-content {
+  flex-grow: 1;
+  display: flex;
+  align-items: flex-start;
+}
+
+.row-side-panel {
+  flex-grow: 0;
+}
+
+.col {
+  float: left;
+  position: relative;
+  min-height: 1px;
+  padding: 0 15px;
+  max-width: 400px;
+}
