+++ static/css/modal.css
@@ -1,0 +1,94 @@
+/* ===== MODAL ===== */
+.modal {
+  display: flex;
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.5);
+  justify-content: center;
+  align-items: center;
+}
+
+#overlay-container {
+  position: fixed;
+  width: 100%;
+  z-index: 1000;
+}
+
+/* ===== SETTINGS PAGE ===== */
+.settings-page {
+  position: relative;
+  width: 100%;
+  max-width: 500px;
+  margin: 20px;
+  padding: 20px;
+  background-color: var(--background-color);
+  border: 1px solid var(--border-color);
+  border-radius: 8px;
+  box-shadow: 0 4px 20px rgba(0, 0, 0, 0.15);
+}
+
+.settings-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 20px;
+  padding-bottom: 10px;
+  border-bottom: 1px solid var(--border-color);
+}
+
+.settings-group {
+  margin-bottom: 24px;
+}
+
+.settings-group h3 {
+  margin-bottom: 12px;
+}
+
+.setting-item {
+  margin-left: 10px;
+}
+
+.setting-description {
+  font-size: 0.9em;
+  color: var(--footer-link-color);
+  margin-top: 5px;
+  margin-bottom: 15px;
+}
+
+.radio-group {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+}
+
+.settings-page input[type="checkbox"],
+.settings-page input[type="radio"] {
+  margin-right: 8px;
+  cursor: pointer;
+  width: 16px;
+  height: 16px;
+  vertical-align: middle;
+}
+
+.settings-page label {
+  display: flex;
+  align-items: center;
+  cursor: pointer;
+  margin-bottom: 5px;
+}
+
+.settings-page h2 {
+  color: var(--text-color);
+  font-size: 1.5em;
+  font-weight: 500;
+  margin-top: 0;
+}
+
+.settings-page h3 {
+  color: var(--text-color);
+  font-size: 1.2em;
+  font-weight: 500;
+}
