--- static/css/tree-list.css
@@ -1,65 +1,0 @@
-
-.tree-wrapper  {
-  padding-top: 10px;
-}
-.tree-list {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-}
-.tree-list .tree-item {
-  position: relative;
-  display: block;
-  min-height: 2em;
-  line-height: 2em;
-  margin-bottom: 10px;
-  padding-left: 21px;
-}
-.tree-list .tree-item:before, .tree-list .tree-item:after {
-  content: '';
-  position: absolute;
-  display: block;
-  background-color: light-dark(#d1d5da,#313438);
-}
-.tree-list .tree-item:before {
-  top: 0;
-  left: 10px;
-  width: 1px;
-  height: calc(100% + 10px);
-}
-.tree-list .tree-item:after {
-  top: 1em;
-  left: 10px;
-  width: 11px;
-  height: 1px;
-}
-.tree-list .tree-item:last-child:not(:first-child) {
-  margin-bottom: 0;
-}
-.tree-list .tree-item:last-child:before {
-  height: 1em;
-}
-.tree-list .tree-item:first-child:before {
-  top: -10px;
-  /* height: calc(50% + 10px); */
-  height: 25px;
-}
-.tree-list .tree-item:first-child:not(:last-child):before {
-  top: -10px;
-  height: calc(100% + 20px);
-}
-.tree-list .tree-item:last-child:not(:first-child):before {
-  height: 1em;
-}
-.tree-list .tree-item > span {
-  display: inline-block;
-  padding: 0 5px;
-  border: 1px solid light-dark(#d1d5da,#313438);
-}
-.tree-list .tree-item > span:hover {
-  background-color: lightgray;
-}
-.tree-list .tree-item > .tree-list {
-  padding-top: 10px;
-}
-
