+++ static/index.html
@@ -2,10 +2,14 @@
 <html>
 <head>
 	<title>New Tab</title>
-	<link href="./css/custom.css" rel="stylesheet" />
-	<link href="./css/tree-list.css" rel="stylesheet" />
-	<!-- <script src="./lib/van-1.5.3.nomodule.js"></script> -->
-	<!-- <script src="./lib/van-1.5.3.nomodule.debug.js"></script> -->
+	<link href="./css/variables.css" rel="stylesheet" />
+	<link href="./css/base.css" rel="stylesheet" />
+	<link href="./css/layout.css" rel="stylesheet" />
+	<link href="./css/buttons.css" rel="stylesheet" />
+	<link href="./css/tree.css" rel="stylesheet" />
+	<link href="./css/modal.css" rel="stylesheet" />
+	<link href="./css/forms.css" rel="stylesheet" />
+	<link href="./css/footer.css" rel="stylesheet" />
 </head>
 <body>
 	<script src="../dist/app.js"></script>
