--- ts/constants.ts
@@ -1,52 +1,0 @@
-import { FaviconProvider, Settings } from './types';
-
-// Constants
-export const CURRENT_LIST_VERSION = 'links-v1';
-export const SETTINGS_VERSION = 'settings-v1';
-export const DOM_CLASSES = {
-  DISPLAY_NONE: 'display-none',
-  TREE_LIST: 'tree-list',
-  TREE_ITEM: 'tree-item',
-  TEXT_PARENT: 'text-parent',
-  TEXT_CHILD: 'text-child',
-  TEXT_LINETHROUGH: 'text-linethrough',
-  WELCOME_MESSAGE: 'welcome-message',
-  SETTINGS_PAGE: 'settings-page',
-  BUTTON_GROUP: 'button-group'
-};
-
-export const ICONS = {
-  MINUS: '[\u2212]',
-  PLUS: '[+]',
-  UP: '↑',
-  DOWN: '↓',
-  EDIT: `<svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
-    <path d="M17 3a2.85 2.83 0 1 1 4 4L7.5 20.5 2 22l1.5-5.5L17 3z"></path>
-  </svg>`,
-  CLOSE: `<svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
-    <line x1="18" y1="6" x2="6" y2="18"></line>
-    <line x1="6" y1="6" x2="18" y2="18"></line>
-  </svg>`,
-  SETTINGS: `<svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
-    <circle cx="12" cy="12" r="3"></circle>
-    <path d="M19.4 15a1.65 1.65 0 0 0 .33 1.82l.06.06a2 2 0 0 1 0 2.83 2 2 0 0 1-2.83 0l-.06-.06a1.65 1.65 0 0 0-1.82-.33 1.65 1.65 0 0 0-1 1.51V21a2 2 0 0 1-2 2 2 2 0 0 1-2-2v-.09A1.65 1.65 0 0 0 9 19.4a1.65 1.65 0 0 0-1.82.33l-.06.06a2 2 0 0 1-2.83 0 2 2 0 0 1 0-2.83l.06-.06a1.65 1.65 0 0 0 .33-1.82 1.65 1.65 0 0 0-1.51-1H3a2 2 0 0 1-2-2 2 2 0 0 1 2-2h.09A1.65 1.65 0 0 0 4.6 9a1.65 1.65 0 0 0-.33-1.82l-.06-.06a2 2 0 0 1 0-2.83 2 2 0 0 1 2.83 0l.06.06a1.65 1.65 0 0 0 1.82.33H9a1.65 1.65 0 0 0 1-1.51V3a2 2 0 0 1 2-2 2 2 0 0 1 2 2v.09a1.65 1.65 0 0 0 1 1.51 1.65 1.65 0 0 0 1.82-.33l.06-.06a2 2 0 0 1 2.83 0 2 2 0 0 1 0 2.83l-.06.06a1.65 1.65 0 0 0-.33 1.82V9a1.65 1.65 0 0 0 1.51 1H21a2 2 0 0 1 2 2 2 2 0 0 1-2 2h-.09a1.65 1.65 0 0 0-1.51 1z"></path>
-  </svg>`,
-  LINK: `<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
-    <path d="M10 13a5 5 0 0 0 7.54.54l3-3a5 5 0 0 0-7.07-7.07l-1.72 1.71"></path>
-    <path d="M14 11a5 5 0 0 0-7.54-.54l-3 3a5 5 0 0 0 7.07 7.07l1.71-1.71"></path>
-  </svg>`
-};
-
-export const FAVICON_PROVIDER_NAMES: { [key in FaviconProvider]: string } = {
-  [FaviconProvider.Chrome]: 'Dynamic (Chrome cache)',
-  [FaviconProvider.DuckDuckGo]: 'Dynamic (DuckDuckGo API)',
-  [FaviconProvider.Generic]: 'Generic link icon',
-  [FaviconProvider.None]: 'None (disabled)',
-}
-
-export const DEFAULT_SETTINGS: Settings = {
-  defaultFaviconProvider: Object.keys(FAVICON_PROVIDER_NAMES)[0] as FaviconProvider,
-  enableRightClickComplete: true,
-  theme: 'system'
-}
-
