+++ ts/services/storage.service.ts
@@ -1,16 +1,10 @@
-import { AppState } from './app.state';
-import { CURRENT_LIST_VERSION, DEFAULT_SETTINGS, SETTINGS_VERSION } from './constants';
-import { FaviconProvider, LinkNodeFlat, Settings } from './types';
+import { CURRENT_LIST_VERSION, DEFAULT_SETTINGS, SETTINGS_VERSION } from '../core/constants';
+import { FaviconProvider, LinkNodeFlat, Settings } from '../core/types';
 
-// Storage Service
 export class StorageService {
-  // Cache for the last saved list to avoid unnecessary storage operations
   private static lastSavedListJson: string = '';
   private static lastSavedSettingsJson: string = '';
 
-  /**
-   * Change this to use max bytes after migrating away from single list
-   */
   static async printStartupInfo() {
     const bytesInUse = await chrome.storage.sync.getBytesInUse("links-v1");
     const maxBytes = chrome.storage.sync.QUOTA_BYTES_PER_ITEM;
@@ -18,17 +12,12 @@ export class StorageService {
   }
 
   static save(list: LinkNodeFlat[]): Promise<void> {
-    // Clone to break references and clean
     const cloneList = JSON.parse(JSON.stringify(list)) as LinkNodeFlat[];
-    
-    // Remove any transient properties
     cloneList.forEach((n: any) => delete n.children);
 
-    // Check if the list has actually changed to avoid unnecessary storage operations
     const listJson = JSON.stringify(cloneList);
     if (this.lastSavedListJson === listJson) {
-      console.log('No changes, skipping save operation');
-      return Promise.resolve(); // No changes, skip save operation
+      return Promise.resolve();
     }
 
     return new Promise((resolve, reject) => {
@@ -46,16 +35,12 @@ export class StorageService {
   }
 
   static async load(): Promise<LinkNodeFlat[]> {
-    const result = await chrome.storage.sync.get(CURRENT_LIST_VERSION)
+    const result = await chrome.storage.sync.get(CURRENT_LIST_VERSION);
     const data = result[CURRENT_LIST_VERSION] as LinkNodeFlat[] | undefined;
     if (data) {
-      // If data is found in storage, use it
       this.lastSavedListJson = JSON.stringify(data);
       return data;
     } else {
-      // If no data is found, this must be a first-time installation.
-      // Load initial data from a static JSON file. This is the new "Welcome" screen.
-      // NOTE: Keep this json file up to date with the latest data format.
       console.log("No data found, loading initial-data JSON file instead...");
       const url = chrome.runtime.getURL('/static/json/initial-data-2.0.0.json');
       const response = await fetch(url);
@@ -66,11 +51,9 @@ export class StorageService {
   }
 
   static saveSettings(settings: Settings): Promise<void> {
-    // Check if the settings have actually changed to avoid unnecessary storage operations
     const settingsJson = JSON.stringify(settings);
     if (this.lastSavedSettingsJson === settingsJson) {
-      console.log('No changes to settings, skipping save operation');
-      return Promise.resolve(); // No changes, skip save operation
+      return Promise.resolve();
     }
 
     return new Promise((resolve, reject) => {
@@ -92,18 +75,16 @@ export class StorageService {
     const data = result[SETTINGS_VERSION] as Settings | undefined;
 
     const mergedSettings = data ? { DEFAULT_SETTINGS, ...data } : DEFAULT_SETTINGS;
-    
+
     if (mergedSettings) {
       this.lastSavedSettingsJson = JSON.stringify(mergedSettings);
     }
-    
+
     return mergedSettings;
   }
 
   static applyNodeDefaults(nodes: LinkNodeFlat[]) {
-    console.log('Parsing node list', nodes);
     nodes.forEach((node) => {
-      // Favicon-related options
       if (node.url) {
         if (node.border === undefined) {
           node.border = 1;
@@ -112,12 +93,10 @@ export class StorageService {
           node.icon = FaviconProvider.Chrome;
         }
       } else {
-        // Cleanup
         delete node.border;
         delete node.icon;
       }
       if (!node.taskComplete) {
-        // Cleanup
         delete node.taskComplete;
       }
     });
