+++ ts/services/validator.service.ts
@@ -1,0 +1,5 @@
+export class ValidatorService {
+  static isValidUrl(url: string): boolean {
+    return /^https?:\/\//.test(url);
+  }
+}
